Modified the other help commands
