+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/help/HelpManager.java
@@ -23,23 +23,49 @@
  */
 package com.blackducksoftware.integration.hub.detect.help;
 
+import java.io.Console;
+import java.io.PrintStream;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.blackducksoftware.integration.hub.detect.DetectConfiguration;
+import com.blackducksoftware.integration.hub.detect.help.print.HelpHtmlWriter;
 import com.blackducksoftware.integration.hub.detect.help.print.HelpPrinter;
+import com.blackducksoftware.integration.hub.detect.interactive.InteractiveManager;
+import com.blackducksoftware.integration.hub.detect.interactive.reader.ConsoleInteractiveReader;
+import com.blackducksoftware.integration.hub.detect.interactive.reader.InteractiveReader;
+import com.blackducksoftware.integration.hub.detect.interactive.reader.ScannerInteractiveReader;
 
 @Component
 public class HelpManager {
-    Map<String, List<String>> commandToAction;
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    @Autowired
+    HelpHtmlWriter helpHtmlWriter;
+
+    @Autowired
+    InteractiveManager interactiveManager;
 
     public boolean isHelpMessageApplicable(final String[] applicationArgs) {
-        final HelpState helpState = new HelpState(applicationArgs);
-        return helpState.isHelpMessage();
+        return isCommand(applicationArgs, "-h", "--help");
+    }
+
+    public boolean isHelpDocumentApplicable(final String[] applicationArgs) {
+        return isCommand(applicationArgs, "-hdoc", "--helpdocument");
+    }
+
+    public boolean isInteractiveModeApplicable(final String[] applicationArgs) {
+        return isCommand(applicationArgs, "-i", "--interactive");
+    }
+
+    private boolean isCommand(final String[] applicationArgs, final String command, final String longCommand) {
+        return applicationArgs.length >= 1 && (command.equals(applicationArgs[0]) || longCommand.equals(applicationArgs[0]));
     }
 
     public void printAppropriateHelpMessage(final String[] applicationArgs, final DetectOptionManager detectOptionManager) {
@@ -75,6 +101,26 @@ public class HelpManager {
                     .filter(option -> DetectConfiguration.GROUP_HUB_CONFIGURATION.equals(option.getGroup()))
                     .collect(Collectors.toList());
             helpPrinter.printHelpMessage(filteredCommonOptions);
+        }
+    }
+
+    public void writeHelpMessage(final String detectVersion) {
+        helpHtmlWriter.writeHelpMessage(String.format("hub-detect-%s-help.html", detectVersion));
+    }
+
+    public void runInteractiveMode() {
+        final InteractiveReader interactiveReader = createInteractiveReader();
+        final PrintStream interactivePrintStream = new PrintStream(System.out);
+        interactiveManager.interact(interactiveReader, interactivePrintStream);
+    }
+
+    private InteractiveReader createInteractiveReader() {
+        final Console console = System.console();
+        if (console != null) {
+            return new ConsoleInteractiveReader(console);
+        } else {
+            logger.warn("It may be insecure to enter passwords because you are running in a virtual console.");
+            return new ScannerInteractiveReader(System.in);
         }
     }
 }
